IGRFBuilderWithGeocentricComponents
